Fix weibull_pdf, broken for R >= 4.2.0
This function was calling if() with a condition of length >1, which was incorrect,
because the logic would then only be based on the first element of the vector. For
the most part this did not affect operation because none of the elements in the vector
produced by dweibull were NA.

However, starting in R 4.2.0, this usage of if results in an error.

This changes replaces if with the corrected vectorized ifelse, which fixes the call
and unbreaks the weibull distribution for R 4.2.0.